Update README.md
This repository contains a simple network port scanner implemented in Python. The scanner allows users to specify one or multiple target IP addresses and the number of ports to scan. It iterates through the specified ports and checks if they are open on the target IP addresses. The program utilizes socket programming to establish connections with the specified ports. Additionally, it provides basic error handling and colorized output using the termcolor library.